feat: Refactor findSolution.js and isEven.js
The code changes refactor the findSolution.js and isEven.js files. In findSolution.js, the console.log statements are added to provide debugging information. In isEven.js, the even variable is initialized to 0 before checking if the number is even. These changes improve the readability and functionality of the code.